requirements.txt
requests
pillow
pytz
tqdm

